Fix translation and move some files
