+++ frontend/pages/signup/index.tsx
@@ -1,7 +1,7 @@
 import Page1 from "./formPages/Page1";
 import Page2 from "./formPages/Page2";
-import FormController from "./FormController";
-import { FormSchema, FormFields } from "./FormSchema";
+import FormController from "../../components/FormController";
+import { SignupFormSchema, SignupFormFields } from "./SignupFormSchema";
 import { Box } from "@mui/system";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -9,9 +9,9 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { NextPageContext } from "next";
 
 export default function SignUp() {
-  const form = useForm<FormFields>({
-    resolver: zodResolver(FormSchema),
-    mode: "onChange",
+  const form = useForm<SignupFormFields>({
+    resolver: zodResolver(SignupFormSchema),
+    mode: "onTouched",
   });
 
   return (
